employee-portal/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
redis==5.0.3
pyjwt==2.8.0
bcrypt==4.1.2
boto3==1.34.74
pika==1.3.2
sendgrid==6.11.0
python-multipart==0.0.9
pydantic==2.6.4
